Made conveyor ignore angular velocity
There was a bug with the cracker box, that made it so when the object
fell into the conveyor, it gained some angular velocity. This caused
the box to spin while moving on the conveyor, thus would have caused us
not to know where to grip it in the future. As a fix, I made the
conveyor turn off angular velocity as well.